# server-go/docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateDatasourceRequest:
    properties:
      config:
        $ref: '#/definitions/models.DatasourceConfig'
      description:
        type: string
      mask_secrets:
        description: 'If true, secrets are masked in API responses (default: true)'
        type: boolean
      name:
        type: string
      supported_schemas:
        items:
          type: string
        type: array
      tags:
        items:
          type: string
        type: array
      type:
        allOf:
        - $ref: '#/definitions/models.DatasourceType'
        description: 'LEGACY: Keep for backwards compatibility'
      type_config:
        additionalProperties: true
        type: object
      type_id:
        description: 'NEW: Registry-based type system (preferred)'
        type: string
    required:
    - name
    type: object
...
  models.Datasource:
    properties:
      config:
        $ref: '#/definitions/models.DatasourceConfig'
      created_at:
        type: string
      description:
        type: string
      health:
        $ref: '#/definitions/models.HealthInfo'
      id:
        type: string
      mask_secrets:
        description: If true, secrets are masked in API responses
        type: boolean
      name:
        type: string
      supported_schemas:
        description: Control schema support - which control schemas this connection
          supports
        items:
          type: string
        type: array
      tags:
        items:
          type: string
        type: array
      type:
        allOf:
        - $ref: '#/definitions/models.DatasourceType'
        description: 'LEGACY: Keep for backwards compatibility during migration'
      type_config:
        additionalProperties: true
        type: object
      type_id:
        description: |-
          NEW: Registry-based type system (preferred)
          TypeID format: "category.name" (e.g., "db.postgres", "stream.websocket-bidir")
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
  models.DatasourceConfig:
    properties:
      api:
        $ref: '#/definitions/models.APIConfig'
      csv:
        $ref: '#/definitions/models.CSVConfig'
      edgelake:
        $ref: '#/definitions/models.EdgeLakeConfig'
      frigate:
        $ref: '#/definitions/models.FrigateConfig'
      mqtt:
        $ref: '#/definitions/models.MQTTConfig'
      prometheus:
        $ref: '#/definitions/models.PrometheusConfig'
      socket:
        $ref: '#/definitions/models.SocketConfig'
      sql:
        $ref: '#/definitions/models.SQLConfig'
      tsstore:
        $ref: '#/definitions/models.TSStoreConfig'
    type: object
  models.DatasourceType:
    enum:
    - sql
    - csv
    - socket
    - api
    - tsstore
    - prometheus
    - edgelake
    - mqtt
    - frigate
    type: string
    x-enum-varnames:
    - DatasourceTypeSQL
    - DatasourceTypeCSV
    - DatasourceTypeSocket
    - DatasourceTypeAPI
    - DatasourceTypeTSStore
    - DatasourceTypePrometheus
    - DatasourceTypeEdgeLake
    - DatasourceTypeMQTT
    - DatasourceTypeFrigate
...
  models.HealthInfo:
    properties:
      error_message:
        type: string
      last_check:
        type: string
      last_success:
        type: string
      response_time:
        description: milliseconds
        type: integer
      status:
        $ref: '#/definitions/models.HealthStatus'
    type: object
...
  models.Query:
    properties:
      params:
        additionalProperties: true
        description: Query parameters
        type: object
      raw:
        description: Raw query string (SQL, filter expression, etc.)
        type: string
      type:
        allOf:
        - $ref: '#/definitions/models.QueryType'
        description: Query type
    type: object
  models.QueryRequest:
    properties:
      query:
        $ref: '#/definitions/models.Query'
    required:
    - query
    type: object
  models.QueryResponse:
    properties:
      duration:
        description: milliseconds
        type: integer
      error:
        type: string
      result_set:
        $ref: '#/definitions/models.ResultSet'
      success:
        type: boolean
    type: object
...
  models.ResultSet:
    properties:
      columns:
        description: Column names
        items:
          type: string
        type: array
      metadata:
        additionalProperties: true
        description: Additional metadata
        type: object
      rows:
        description: Data rows
        items:
          items: {}
          type: array
        type: array
    type: object
...
  models.SchemaResponse:
    properties:
      duration:
        description: milliseconds
        type: integer
      error:
        type: string
      prometheus_schema:
        allOf:
        - $ref: '#/definitions/models.PrometheusSchemaInfo'
        description: For Prometheus datasources
      schema:
        allOf:
        - $ref: '#/definitions/models.SchemaInfo'
        description: For SQL datasources
      success:
        type: boolean
    type: object
...
  models.TestDatasourceRequest:
    properties:
      config:
        $ref: '#/definitions/models.DatasourceConfig'
      id:
        description: 'Optional: existing connection ID to resolve masked secrets from
          DB'
        type: string
      type:
        allOf:
        - $ref: '#/definitions/models.DatasourceType'
        description: 'LEGACY: Keep for backwards compatibility'
      type_config:
        additionalProperties: true
        type: object
      type_id:
        description: 'NEW: Registry-based type system (preferred)'
        type: string
    type: object
  models.TestDatasourceResponse:
    properties:
      data: {}
      message:
        type: string
      response_time:
        description: milliseconds
        type: integer
      status:
        $ref: '#/definitions/models.HealthStatus'
      success:
        type: boolean
    type: object
...
  models.UpdateDatasourceRequest:
    properties:
      config:
        allOf:
        - $ref: '#/definitions/models.DatasourceConfig'
        description: 'LEGACY: Keep for backwards compatibility'
      description:
        type: string
      mask_secrets:
        description: If provided, updates secret masking setting
        type: boolean
      name:
        type: string
      supported_schemas:
        items:
          type: string
        type: array
      tags:
        items:
          type: string
        type: array
      type_config:
        additionalProperties: true
        type: object
      type_id:
        description: 'NEW: Registry-based type system'
        type: string
    type: object
...
host: localhost:3001
info:
  contact:
    email: support@example.com
    name: Dashboard Team
  description: Dashboard system with AI-powered chart generation
  title: TRVE Dashboards API
  version: "1.0"
paths:
...
  /connections/{id}/mqtt/sample:
    get:
      description: Subscribe to a topic and return the first message's schema (columns
        and sample values)
      parameters:
      - description: Datasource ID
        in: path
        name: id
        required: true
        type: string
      - description: MQTT topic to sample
        in: query
        name: topic
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Sample a single MQTT topic
      tags:
      - datasources
  /connections/{id}/mqtt/topics:
    get:
      description: Subscribe briefly to discover available topics on an MQTT broker
      parameters:
      - description: Datasource ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get topics from an MQTT broker
      tags:
      - datasources
...
  /datasources:
    get:
      description: Retrieve all datasources with pagination and optional type/tag
        filters
      parameters:
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      - description: Filter by datasource type (api, websocket, file)
        in: query
        name: type
        type: string
      - collectionFormat: csv
        description: Filter by tags (OR semantics, repeat param)
        in: query
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List all datasources
      tags:
      - datasources
    post:
      consumes:
      - application/json
      description: Create a new data source (API, WebSocket, or File)
      parameters:
      - description: Datasource to create
        in: body
        name: datasource
        required: true
        schema:
          $ref: '#/definitions/models.CreateDatasourceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Datasource'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new datasource
      tags:
      - datasources
  /datasources/{id}:
    delete:
      description: Delete a datasource by ID
      parameters:
      - description: Datasource ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Delete a datasource
      tags:
      - datasources
    get:
      description: Retrieve a single datasource by its ID
      parameters:
      - description: Datasource ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Datasource'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a datasource by ID
      tags:
      - datasources
    put:
      consumes:
      - application/json
      description: Update an existing datasource by ID
      parameters:
      - description: Datasource ID
        in: path
        name: id
        required: true
        type: string
      - description: Datasource updates
        in: body
        name: datasource
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDatasourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Datasource'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a datasource
      tags:
      - datasources
  /datasources/{id}/edgelake/databases:
    get:
      description: Retrieve all database names from an EdgeLake node's blockchain
        registry
      parameters:
      - description: Datasource ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get databases from an EdgeLake data source
      tags:
      - datasources
  /datasources/{id}/edgelake/schema:
    get:
      description: Retrieve column names and types for a specific table from an EdgeLake
        node
      parameters:
      - description: Datasource ID
        in: path
        name: id
        required: true
        type: string
      - description: Database name
        in: query
        name: database
        required: true
        type: string
      - description: Table name
        in: query
        name: table
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get table schema from an EdgeLake data source
      tags:
      - datasources
  /datasources/{id}/edgelake/tables:
    get:
      description: Retrieve table names for a specific database from an EdgeLake node
      parameters:
      - description: Datasource ID
        in: path
        name: id
        required: true
        type: string
      - description: Database name
        in: query
        name: database
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get tables from an EdgeLake data source
      tags:
      - datasources
...
  /datasources/{id}/prometheus/labels/{label}/values:
    get:
      description: Retrieve all possible values for a specific label from a Prometheus
        datasource
      parameters:
      - description: Datasource ID
        in: path
        name: id
        required: true
        type: string
      - description: Label name
        in: path
        name: label
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get values for a Prometheus label
      tags:
      - datasources
  /datasources/{id}/query:
    post:
      consumes:
      - application/json
      description: Execute a query and return normalized results
      parameters:
      - description: Datasource ID
        in: path
        name: id
        required: true
        type: string
      - description: Query to execute
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/models.QueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QueryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Execute a query against a datasource
      tags:
      - datasources
  /datasources/{id}/schema:
    get:
      description: Retrieve tables and columns for SQL datasources. Only SQL-type
        datasources support this endpoint.
      parameters:
      - description: Datasource ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SchemaResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get database schema for a SQL datasource
      tags:
      - datasources
...
  /datasources/test:
    post:
      consumes:
      - application/json
      description: Test a datasource connection without saving it
      parameters:
      - description: Datasource configuration to test
        in: body
        name: datasource
        required: true
        schema:
          $ref: '#/definitions/models.TestDatasourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TestDatasourceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Test a datasource connection
      tags:
      - datasources
...
swagger: "2.0"